refactor: moved and rename entiies columns
